refactor: separate editor, link dialog and show loading
