+++ lib/app/modules/notices/presentation/pages/notice_write_body_page.dart
@@ -5,7 +5,6 @@ import 'package:dotted_border/dotted_border.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:flutter_quill/extensions.dart';
 import 'package:flutter_quill/flutter_quill.dart';
 import 'package:image_picker/image_picker.dart';
 import 'package:keyboard_actions/keyboard_actions.dart';
@@ -14,11 +13,12 @@ import 'package:ziggle/app/modules/common/presentation/extensions/toast.dart';
 import 'package:ziggle/app/modules/common/presentation/widgets/ziggle_app_bar.dart';
 import 'package:ziggle/app/modules/common/presentation/widgets/ziggle_back_button.dart';
 import 'package:ziggle/app/modules/common/presentation/widgets/ziggle_button.dart';
-import 'package:ziggle/app/modules/common/presentation/widgets/ziggle_input.dart';
 import 'package:ziggle/app/modules/notices/presentation/bloc/ai_bloc.dart';
 import 'package:ziggle/app/modules/notices/presentation/bloc/notice_write_bloc.dart';
 import 'package:ziggle/app/modules/notices/presentation/extensions/quill.dart';
+import 'package:ziggle/app/modules/notices/presentation/widgets/editor.dart';
 import 'package:ziggle/app/modules/notices/presentation/widgets/language_toggle.dart';
+import 'package:ziggle/app/modules/notices/presentation/widgets/link_dialog.dart';
 import 'package:ziggle/app/modules/notices/presentation/widgets/photo_item.dart';
 import 'package:ziggle/app/router.gr.dart';
 import 'package:ziggle/app/values/palette.dart';
@@ -179,45 +179,7 @@ class _LayoutState extends State<_Layout> with SingleTickerProviderStateMixin {
                 ],
               ),
             ),
-            Expanded(
-              child: TabBarView(
-                controller: _tabController,
-                physics: const NeverScrollableScrollPhysics(),
-                children: [
-                  _Editor(
-                    titleFocusNode: _koreanTitleFocusNode,
-                    bodyFocusNode: _koreanBodyFocusNode,
-                    titleController: _koreanTitleController,
-                    bodyController: _koreanBodyController,
-                  ),
-                  _Editor(
-                    onTranslate: _englishBodyController
-                            .plainTextEditingValue.text
-                            .trim()
-                            .isNotEmpty
-                        ? null
-                        : () async {
-                            final bloc = context.read<AiBloc>();
-                            final blocker =
-                                bloc.stream.firstWhere((s) => s.hasResult);
-                            bloc.add(AiEvent.request(
-                              body: _koreanBodyController.html,
-                              lang: AppLocale.en,
-                            ));
-                            final result = await blocker;
-                            result.mapOrNull(
-                              loaded: (result) =>
-                                  _englishBodyController.html = result.body,
-                            );
-                          },
-                    titleFocusNode: _englishTitleFocusNode,
-                    bodyFocusNode: _englishBodyFocusNode,
-                    titleController: _englishTitleController,
-                    bodyController: _englishBodyController,
-                  ),
-                ],
-              ),
-            ),
+            _buildEditors(),
             if (!_koreanTitleFocusNode.hasFocus &&
                 !_koreanBodyFocusNode.hasFocus &&
                 !_englishTitleFocusNode.hasFocus &&
@@ -285,6 +247,50 @@ class _LayoutState extends State<_Layout> with SingleTickerProviderStateMixin {
     );
   }
 
+  Expanded _buildEditors() {
+    return Expanded(
+      child: TabBarView(
+        controller: _tabController,
+        physics: const NeverScrollableScrollPhysics(),
+        children: [
+          Editor(
+            titleFocusNode: _koreanTitleFocusNode,
+            bodyFocusNode: _koreanBodyFocusNode,
+            titleController: _koreanTitleController,
+            bodyController: _koreanBodyController,
+          ),
+          BlocBuilder<AiBloc, AiState>(
+            builder: (context, state) => Editor(
+              translating: state.isLoading,
+              onTranslate: _englishBodyController.plainTextEditingValue.text
+                      .trim()
+                      .isNotEmpty
+                  ? null
+                  : _translate,
+              titleFocusNode: _englishTitleFocusNode,
+              bodyFocusNode: _englishBodyFocusNode,
+              titleController: _englishTitleController,
+              bodyController: _englishBodyController,
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+
+  Future<void> _translate() async {
+    final bloc = context.read<AiBloc>();
+    final blocker = bloc.stream.firstWhere((s) => s.hasResult);
+    bloc.add(AiEvent.request(
+      body: _koreanBodyController.html,
+      lang: AppLocale.en,
+    ));
+    final result = await blocker;
+    result.mapOrNull(
+      loaded: (result) => _englishBodyController.html = result.body,
+    );
+  }
+
   List<ButtonBuilder> _buildToolbarButtons(QuillController controller) {
     return [
       (_) => _buildToggleButton(
@@ -315,7 +321,7 @@ class _LayoutState extends State<_Layout> with SingleTickerProviderStateMixin {
                   showCupertinoDialog<QuillTextLink>(
                     context: context,
                     barrierDismissible: true,
-                    builder: (context) => _LinkDialog(controller: controller),
+                    builder: (context) => LinkDialog(controller: controller),
                   ).then((link) => link?.submit(controller));
                   options.afterButtonPressed?.call();
                 },
@@ -370,152 +376,4 @@ class _LayoutState extends State<_Layout> with SingleTickerProviderStateMixin {
           child: child,
         ),
       );
-}
-
-class _Editor extends StatelessWidget {
-  const _Editor({
-    required this.titleFocusNode,
-    required this.bodyFocusNode,
-    required this.titleController,
-    required this.bodyController,
-    this.onTranslate,
-  });
-
-  final FocusNode titleFocusNode;
-  final FocusNode bodyFocusNode;
-  final TextEditingController titleController;
-  final QuillController bodyController;
-  final VoidCallback? onTranslate;
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      children: [
-        Padding(
-          padding: const EdgeInsets.symmetric(horizontal: 18),
-          child: ZiggleInput(
-            controller: titleController,
-            focusNode: titleFocusNode,
-            showBorder: false,
-            hintText: context.t.notice.write.titleHint,
-            style: const TextStyle(
-              fontSize: 24,
-              fontWeight: FontWeight.bold,
-            ),
-          ),
-        ),
-        const SizedBox(height: 10),
-        Padding(
-          padding: const EdgeInsets.symmetric(horizontal: 18),
-          child: Container(height: 1, color: Palette.grayBorder),
-        ),
-        const SizedBox(height: 10),
-        if (onTranslate != null)
-          Padding(
-            padding: const EdgeInsets.symmetric(horizontal: 18),
-            child: Row(
-              children: [
-                ZiggleButton.big(
-                  emphasize: false,
-                  onPressed: onTranslate,
-                  child: Row(
-                    children: [
-                      Assets.icons.sparks.svg(),
-                      const SizedBox(width: 10),
-                      Text(context.t.notice.write.translate),
-                    ],
-                  ),
-                ),
-              ],
-            ),
-          ),
-        Expanded(
-          child: Padding(
-            padding: const EdgeInsets.symmetric(horizontal: 18),
-            child: QuillEditor.basic(
-              focusNode: bodyFocusNode,
-              controller: bodyController,
-              configurations: QuillEditorConfigurations(
-                placeholder: context.t.notice.write.bodyHint,
-                padding: const EdgeInsets.symmetric(vertical: 10),
-                customStyles: const DefaultStyles(
-                  placeHolder: DefaultTextBlockStyle(
-                    TextStyle(fontSize: 16, color: Palette.gray),
-                    HorizontalSpacing.zero,
-                    VerticalSpacing.zero,
-                    VerticalSpacing.zero,
-                    null,
-                  ),
-                ),
-              ),
-            ),
-          ),
-        ),
-      ],
-    );
-  }
-}
-
-class _LinkDialog extends StatefulWidget {
-  const _LinkDialog({
-    required QuillController controller,
-  }) : _controller = controller;
-
-  final QuillController _controller;
-
-  @override
-  State<_LinkDialog> createState() => _LinkDialogState();
-}
-
-class _LinkDialogState extends State<_LinkDialog> {
-  late final _initialTextLink = QuillTextLink.prepare(widget._controller);
-  late final _text = TextEditingController(text: _initialTextLink.text);
-  late final _link = TextEditingController(text: _initialTextLink.link ?? '');
-
-  @override
-  Widget build(BuildContext context) {
-    return CupertinoAlertDialog(
-      content: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          CupertinoTextField(
-            placeholder: context.t.notice.write.link.text,
-            controller: _text,
-            onChanged: (v) => setState(() {}),
-          ),
-          const SizedBox(height: 8),
-          CupertinoTextField(
-            placeholder: context.t.notice.write.link.link,
-            controller: _link,
-            onChanged: (v) => setState(() {}),
-          ),
-        ],
-      ),
-      actions: [
-        CupertinoDialogAction(
-          onPressed: _text.text.isEmpty ||
-                  _link.text.isEmpty ||
-                  !const AutoFormatMultipleLinksRule()
-                      .oneLineLinkRegExp
-                      .hasMatch(_link.text)
-              ? null
-              : () => Navigator.pop(
-                    context,
-                    QuillTextLink(_text.text, _link.text),
-                  ),
-          child: _initialTextLink.link == null
-              ? Text(context.t.notice.write.link.add)
-              : Text(context.t.notice.write.link.change),
-        ),
-        CupertinoDialogAction(
-          isDestructiveAction: true,
-          onPressed: () => Navigator.pop(
-            context,
-            QuillTextLink(_text.text, null),
-          ),
-          child: Text(context.t.notice.write.link.remove),
-        ),
-      ],
-    );
-  }
 }
